refactor(linter): split ambient contracts module
